inventory-app/requirements.txt
# Microservice Inventaire — dépendances Python
Flask==3.0.3
psycopg2-binary==2.9.9
gunicorn==22.0.0
